requirements.txt
requests
python-dotenv
mcp
youtube-search
youtube-transcript-api
googlesearch-python